+++ core/tests/protractor/collections.js
@@ -20,14 +20,21 @@ var general = require('../protractor_utils/general.js');
 var users = require('../protractor_utils/users.js');
 var AdminPage = require('../protractor_utils/AdminPage.js');
 var collectionEditor = require('../protractor_utils/collectionEditor.js');
+var CreatorDashboardPage =
+  require('../protractor_utils/CreatorDashboardPage.js');
 
 
 describe('Collections', function() {
   var adminPage = null;
+  var creatorDashboardPage = null;
+  var COLLECTION_ID = null;
 
   beforeAll(function() {
     adminPage = new AdminPage.AdminPage();
+    creatorDashboardPage = new CreatorDashboardPage.CreatorDashboardPage();
     var USERNAME = 'aliceCollections';
+    var USERNAME_TWO = 'playerCollections';
+    users.createUser('player@collections.com', USERNAME_TWO);
     users.createUser('alice@collections.com', USERNAME);
     users.createAndLoginAdminUser('testadm@collections.com', 'testadm');
     adminPage.get();
@@ -36,6 +43,28 @@ describe('Collections', function() {
     browser.waitForAngular();
     adminPage.reloadAllExplorations();
     adminPage.updateRole(USERNAME, 'collection editor');
+    adminPage.updateRole(USERNAME_TWO, 'collection editor');
+    users.logout();
+
+    users.login('player@collections.com');
+    browser.get(general.SERVER_URL_PREFIX);
+    var dropdown = element(by.css('.protractor-test-profile-dropdown'));
+    browser.actions().mouseMove(dropdown).perform();
+    dropdown.element(by.css('.protractor-test-dashboard-link')).click();
+    browser.waitForAngular();
+    creatorDashboardPage.clickCreateActivityButton();
+    creatorDashboardPage.clickCreateCollectionButton();
+    browser.waitForAngular();
+    // Add existing explorations.
+    collectionEditor.addExistingExploration('0');
+    collectionEditor.saveDraft();
+    collectionEditor.closeSaveModal();
+    collectionEditor.publishCollection();
+    collectionEditor.setTitle('Test Collection 2');
+    collectionEditor.setObjective('This is the second test collection.');
+    collectionEditor.setCategory('Algebra');
+    collectionEditor.saveChanges();
+    browser.waitForAngular();
     users.logout();
   });
 
@@ -46,9 +75,8 @@ describe('Collections', function() {
     browser.actions().mouseMove(dropdown).perform();
     dropdown.element(by.css('.protractor-test-dashboard-link')).click();
     browser.waitForAngular();
-    element(by.css('.protractor-test-create-activity')).click();
-    // Create new collection.
-    element(by.css('.protractor-test-create-collection')).click();
+    creatorDashboardPage.clickCreateActivityButton();
+    creatorDashboardPage.clickCreateCollectionButton();
     browser.waitForAngular();
     // Add existing explorations.
     collectionEditor.addExistingExploration('0');
@@ -82,7 +110,93 @@ describe('Collections', function() {
     users.logout();
   });
 
+  it('checks for console errors in a collection with one node', function() {
+    users.login('player@collections.com');
+    browser.get('/search/find?q=');
+    browser.waitForAngular();
+    element.all(by.css(
+      '.protractor-test-collection-summary-tile-title')
+    ).filter(function (elem) {
+      return elem.getText().then(function (text) {
+        return text === 'Test Collection 2';
+      });
+    }).first().click();
+    browser.waitForAngular();
+    general.waitForSystem();
+    browser.getCurrentUrl().then(function(url) {
+      var pathname = url.split('/');
+      COLLECTION_ID = pathname[4];
+    });
+    users.logout();
+  });
+
+  it('checks for console errors in a collection with two nodes', function() {
+    users.login('player@collections.com');
+    creatorDashboardPage.get();
+    browser.waitForAngular();
+    element.all(by.css(
+      '.protractor-test-collection-card')).first().click();
+    browser.waitForAngular();
+    general.waitForSystem();
+    collectionEditor.addExistingExploration('4');
+    collectionEditor.saveDraft();
+    collectionEditor.setCommitMessage('Add Exploration');
+    collectionEditor.closeSaveModal();
+    browser.waitForAngular();
+
+    browser.get('/collection/' + COLLECTION_ID);
+    browser.waitForAngular();
+    users.logout();
+  });
+
+  it('checks for console errors in a collection with three nodes', function() {
+    users.login('player@collections.com');
+    creatorDashboardPage.get();
+    browser.waitForAngular();
+    element.all(by.css(
+      '.protractor-test-collection-card')).first().click();
+    browser.waitForAngular();
+    general.waitForSystem();
+    collectionEditor.addExistingExploration('13');
+    collectionEditor.saveDraft();
+    collectionEditor.setCommitMessage('Add Exploration');
+    collectionEditor.closeSaveModal();
+    browser.waitForAngular();
+
+    browser.get('/collection/' + COLLECTION_ID);
+    browser.waitForAngular();
+    users.logout();
+  });
+
+  it('checks for console errors in a collection with four nodes', function() {
+    users.login('player@collections.com');
+    creatorDashboardPage.get();
+    browser.waitForAngular();
+    element.all(by.css(
+      '.protractor-test-collection-card')).first().click();
+    browser.waitForAngular();
+    general.waitForSystem();
+    collectionEditor.addExistingExploration('10');
+    collectionEditor.saveDraft();
+    collectionEditor.setCommitMessage('Add Exploration');
+    collectionEditor.closeSaveModal();
+    browser.waitForAngular();
+
+    browser.get('/collection/' + COLLECTION_ID);
+    browser.waitForAngular();
+    users.logout();
+  });
+
   afterEach(function() {
-    general.checkForConsoleErrors([]);
+    browser.manage().logs().get('browser').then(function(browserLog) {
+      var i = 0, severWarnings = false, errors = 0;
+      for (i; i <= browserLog.length - 1; i++) {
+        // Check only for console errors and not warnings
+        if (browserLog[i].level.name_ === 'SEVERE') {
+          errors = errors + 1;
+        }
+      }
+      expect(errors).toBe(0);
+    });
   });
 });
