+++ core/tests/protractor_utils/CreatorDashboardPage.js
@@ -25,6 +25,10 @@ var CreatorDashboardPage = function() {
     element(by.css('.protractor-test-exploration-dashboard-card'));
   var collectionCard = element(by.css('.protractor-test-collection-card'));
   var subscriptionTab = element(by.css('.protractor-test-subscription-tab'));
+  var createActivityButton =
+    element(by.css('.protractor-test-create-activity'));
+  var createCollectionButton =
+    element(by.css('.protractor-test-create-collection'));
 
   this.get = function() {
     return browser.get(CREATOR_DASHBOARD_URL);
@@ -38,6 +42,14 @@ var CreatorDashboardPage = function() {
 
   this.navigateToExplorationEditor = function() {
     explorationDashboardCard.click();
+  };
+
+  this.clickCreateActivityButton = function() {
+    createActivityButton.click();
+  };
+
+  this.clickCreateCollectionButton = function() {
+    createCollectionButton.click();
   };
 
   this.navigateToCollectionEditor = function() {
